Review fixes: - Move category genre base viewset to another file - remove excessive fields in filterset - change title model a little - move user admin to yamdb_user
