+++ api_yamdb/api/filters.py
@@ -1,4 +1,4 @@
-from django_filters import CharFilter, FilterSet, NumberFilter
+from django_filters import CharFilter, FilterSet
 
 from reviews.models import Title
 
@@ -6,8 +6,6 @@ from reviews.models import Title
 class TitleFilter(FilterSet):
     genre = CharFilter(field_name='genre__slug')
     category = CharFilter(field_name='category__slug')
-    year = NumberFilter(field_name='year')
-    name = CharFilter(field_name='name')
 
     class Meta:
         model = Title
